move user.verify package to identity.verify
