+++ src/main/java/io/fusionauth/client/FusionAuthClient.java
@@ -230,12 +230,12 @@ import io.fusionauth.domain.api.user.VerifyEmailRequest;
 import io.fusionauth.domain.api.user.VerifyEmailResponse;
 import io.fusionauth.domain.api.user.VerifyRegistrationRequest;
 import io.fusionauth.domain.api.user.VerifyRegistrationResponse;
-import io.fusionauth.domain.api.user.verify.VerifyCompleteRequest;
-import io.fusionauth.domain.api.user.verify.VerifyCompleteResponse;
-import io.fusionauth.domain.api.user.verify.VerifyRequest;
-import io.fusionauth.domain.api.user.verify.VerifyStartRequest;
-import io.fusionauth.domain.api.user.verify.VerifyStartResponse;
-import io.fusionauth.domain.api.user.verify.VerifySendRequest;
+import io.fusionauth.domain.api.identity.verify.VerifyCompleteRequest;
+import io.fusionauth.domain.api.identity.verify.VerifyCompleteResponse;
+import io.fusionauth.domain.api.identity.verify.VerifyRequest;
+import io.fusionauth.domain.api.identity.verify.VerifyStartRequest;
+import io.fusionauth.domain.api.identity.verify.VerifyStartResponse;
+import io.fusionauth.domain.api.identity.verify.VerifySendRequest;
 import io.fusionauth.domain.oauth2.AccessToken;
 import io.fusionauth.domain.oauth2.DeviceApprovalResponse;
 import io.fusionauth.domain.oauth2.IntrospectResponse;
