+++ src/main/java/io/fusionauth/domain/api/identity/verify/VerifyCompleteRequest.java
@@ -13,14 +13,15 @@
  * either express or implied. See the License for the specific
  * language governing permissions and limitations under the License.
  */
-package io.fusionauth.domain.api.user.verify;
+package io.fusionauth.domain.api.identity.verify;
 
 import io.fusionauth.domain.Buildable;
+import io.fusionauth.domain.api.BaseEventRequest;
 
 /**
- * Verify Send API request object.
+ * Verify Complete API request object.
  */
-public class VerifySendRequest implements Buildable<VerifySendRequest> {
+public class VerifyCompleteRequest extends BaseEventRequest implements Buildable<VerifyCompleteRequest> {
   public String oneTimeCode;
 
   public String verificationId;
