+++ src/main/java/io/fusionauth/domain/api/identity/verify/VerifyStartRequest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2025, FusionAuth, All Rights Reserved
+ * Copyright (c) 2024-2025, FusionAuth, All Rights Reserved
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,16 +13,24 @@
  * either express or implied. See the License for the specific
  * language governing permissions and limitations under the License.
  */
-package io.fusionauth.domain.api.user.verify;
+package io.fusionauth.domain.api.identity.verify;
+
+import java.util.Map;
+import java.util.UUID;
 
 import io.fusionauth.domain.Buildable;
-import io.fusionauth.domain.api.BaseEventRequest;
 
 /**
- * Identity verify request. Used to administratively verify an identity.
+ * @author Brady Wied
  */
-public class VerifyRequest extends BaseEventRequest implements Buildable<VerifyRequest> {
+public class VerifyStartRequest implements Buildable<VerifyStartRequest> {
+  public UUID applicationId;
+
   public String loginId;
 
   public String loginIdType;
+
+  public Map<String, Object> state;
+
+  public String verificationStrategy;
 }
